services/search-agent/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "langchain-community>=0.3.0",
    "tavily-python>=0.5.0",
    "pydantic-settings>=2.6.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
